OLD CODE: not finished work
Some people seem to be interested in the old PoC code for with the raw
PulseAudio API. This is it.

This is work that was not finished, formatted, or anyhthing, and it will be in bad
shape. Please do not comment on the code quality or anything, I know it's
not good, since it was somethng quick I knocked up to see if the
PulseAudio API could be integrated into SuperCollider, which it proved.